Fix regression: Tower BG lerps in reverse direction
This is the same as commit 70357a65bf06a25b7ba98c846ee258ac5eccd88b
("Fix regression: Warp BG lerps in reverse direction"), but for the
tower background.

This bug is most visible when moving the camera in a tower using
invincibility, or holding down ACTION during the credits scroll.