Revert text image fading
This reverts the following commits:

- 29f05c41b1add3aeea3b6b870349020f6cd13b94
- f1bf1f683c459ac4de4de05f938587d196084dea
- a7b22919ae379c96001c8dcdb9f73c36d298f967
- 2ed1aac67d050f19f0a2170ac71f6fe19bb3f747

Recently, text images were changed to fade in with textboxes, where
before they previously appeared after the fade. This created a charming
effect where the images would appear to "load in" once the textbox
finishes fading in. This behavior really complements the retro
aesthetic the game is going for. Changing it to a fade is a needless
change in direction, as it was not a bug in the first place and looked
good already.

Additionally, custom levels have used text images (levelcomplete and
gamecomplete) in creative ways by replacing them with something else
to show as 'foreground' or as a cutscene image. Changing text images
to fade has unintended consequences for levels that have utilized
them in this fashion.