Remove number translations that are numerals
This removes every single translation of a wordy number that just
replaces it with the numeral.

This is because the documentation in README-translators.txt specifically
says

    It's also possible to leave the translations for all the numbers
    empty. In that case, numeric forms will always be used.

However, the translators for Japanese, Korean, and European Portuguese
clearly either didn't read this, or forgot to do so.

There is a very good reason to leave them alone if you want numerals;
namely that if you fill them in, you are prone to making errors. Like,
say, Japanese translating "Twelve" as "23", which is exactly what
happened. By blanking every translation, that error is fixed.